add factory & singleton in services

+++ src/main/java/com/sop/onez/service/ProductService.java
@@ -1,40 +1,56 @@
 package com.sop.onez.service;
 
-import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.sop.onez.model.Product;
 import com.sop.onez.repository.ProductRepository;
 
-@Service
-public class ProductService {
-	
+@Service("productService")
+public class ProductService implements Services {
+
 	@Autowired
 	private ProductRepository productRepository;
-	
-	public List<Product> getAllProduct() {
+
+	@Override
+	public Object getById(int id) {
+		if (productRepository.findById(id).isPresent()) {
+			return productRepository.findById(id).get();
+		}
+		return null;
+	}
+
+	@Override
+	public Object getAll() {
 		return productRepository.findAll();
 	}
 
-	public Product getProduct(int id) {
-		return productRepository.findById(id).get();
-	}
-
-	public void saveProduct(Product product) {
-		int id = ((int) productRepository.count()) + 1;
-		product.setId(id);
+	@Override
+	public void create(Object item) {
+		Product product = (Product) item;
+		product.setId(((int) productRepository.count()) + 1);
 		productRepository.save(product);
 	}
-	
-	public void updateProduct(int id, Product newProduct) {
-		newProduct.setId(id);
-		productRepository.save(newProduct);
+
+	@Override
+	public void update(int id, Object item) {
+		if (productRepository.findById(id).isPresent()) {
+			Product product = (Product) item;
+			product.setId(id);
+			productRepository.save(product);
+		}
 	}
-	
-	public void deleteProduct(int id) {
-		productRepository.deleteById(id);
+
+	@Override
+	public void deleteById(int id) {
+		if (productRepository.findById(id).isPresent()) {
+			productRepository.deleteById(id);
+		}
+	}
+
+	@Override
+	public void deleteAll() {
+		productRepository.deleteAll();
 	}
 
 }
